pawn maybe end

+++ ChessProject/lib/classDir/class_file.py
@@ -47,45 +47,51 @@ class Pawn(Piece):
         list_next_position = sorted(next_position)
         list_next_position[0] = int(list_next_position[0])
         next_position = convert_to_list(next_position)
-        move_list = [abs((next_position[0]) - actual_position[0]), abs((next_position[1]) - (actual_position[1]))]    #deplacement in columns then lines
-        if len(move_list) != 2:                                    #Wrong move
-            print('error: 2 arguments needed')
-        elif move_list[1] > 1 or move_list[0] > 2 or move_list[0] < 1 or move_list[1] < 0:
-            print('error: this move is impossible for a pawn')
-        elif move_list[1] == 1:
-            if move_list[0] == 1:                                  # If this kills another piece - no piece of the same color + piece of the other color needed
-                if chessboard[next_position[0]][list_next_position[1]] == '.':
-                    print('error: this move is impossible for a pawn')
-                elif piece_next_case.color == self.color:           # if that's the same color
-                    print('error: there is already another piece of the same color on this case')
-                elif piece_next_case.color != self.color:              # Other color: it can kill it, the piece which moves take its position and the other disappears
-                    self.position = list_next_position[1] + str(list_next_position[0])
-                    piece_next_case.position = ''
-                    chessboard[list_next_position[0]][list_next_position[1]] = chessboard[list_actual_position[0]][list_actual_position[1]]
-                    chessboard[list_actual_position[0]][list_actual_position[1]] = '.'
-                    self.nb_plays = self.nb_plays + 1                 # The pawn has played, it won't be its first play again
-                    return chessboard
-            else:
-                print('error: this pawn cannot do this move')
-        elif move_list[1] == 0:
-            if chessboard[list_next_position[0]][list_next_position[1]] != '.':  # Next position must be free
-                print('error: there is already another piece there')
-            else:
-                if move_list[0] == 2:
-                    if self.nb_plays == 0:                              # First time this pawn plays, it can advance 2 cases
+        move_list = [next_position[0] - actual_position[0], next_position[1] - actual_position[1]]    #deplacement in columns then lines
+        if self.color == 'white' and move_list[0] > -1:
+            print('error: the pawn cannot move back')
+        elif self.color == 'black' and move_list[0] < 1:
+            print('error: the pawn cannot move back')
+        else:
+            move_list = [abs(move_list[0]), abs(move_list[1])]
+            if len(move_list) != 2:                                    #Wrong move
+                print('error: 2 arguments needed')
+            elif move_list[1] > 1 or move_list[0] > 2 or move_list[0] < 1 or move_list[1] < 0:
+                print('error: this move is impossible for a pawn')
+            elif move_list[1] == 1:
+                if move_list[0] == 1:                                  # If this kills another piece - no piece of the same color + piece of the other color needed
+                    if chessboard[next_position[0]][list_next_position[1]] == '.':
+                        print('error: this move is impossible for a pawn')
+                    elif piece_next_case.color == self.color:           # if that's the same color
+                        print('error: there is already another piece of the same color on this case')
+                    elif piece_next_case.color != self.color:              # Other color: it can kill it, the piece which moves take its position and the other disappears
+                        self.position = list_next_position[1] + str(list_next_position[0])
+                        piece_next_case.position = ''
+                        chessboard[list_next_position[0]][list_next_position[1]] = chessboard[list_actual_position[0]][list_actual_position[1]]
+                        chessboard[list_actual_position[0]][list_actual_position[1]] = '.'
+                        self.nb_plays = self.nb_plays + 1                 # The pawn has played, it won't be its first play again
+                        return chessboard
+                else:
+                    print('error: this pawn cannot do this move')
+            elif move_list[1] == 0:
+                if chessboard[list_next_position[0]][list_next_position[1]] != '.':  # Next position must be free
+                    print('error: there is already another piece there')
+                else:
+                    if move_list[0] == 2:
+                        if self.nb_plays == 0:                              # First time this pawn plays, it can advance 2 cases
+                            self.position = list_next_position[1] + str(list_next_position[0])
+                            chessboard[list_next_position[0]][list_next_position[1]] = chessboard[list_actual_position[0]][list_actual_position[1]]
+                            chessboard[list_actual_position[0]][list_actual_position[1]] = '.'
+                            self.nb_plays = self.nb_plays + 1
+                            return chessboard
+                        elif self.nb_plays > 0:
+                            print('error: this pawn cannot do this move')
+                    if move_list[0] == 1:
                         self.position = list_next_position[1] + str(list_next_position[0])
                         chessboard[list_next_position[0]][list_next_position[1]] = chessboard[list_actual_position[0]][list_actual_position[1]]
                         chessboard[list_actual_position[0]][list_actual_position[1]] = '.'
                         self.nb_plays = self.nb_plays + 1
                         return chessboard
-                    elif self.nb_plays > 0:
-                        print('error: this pawn cannot do this move')
-                if move_list[0] == 1:
-                    self.position = list_next_position[1] + str(list_next_position[0])
-                    chessboard[list_next_position[0]][list_next_position[1]] = chessboard[list_actual_position[0]][list_actual_position[1]]
-                    chessboard[list_actual_position[0]][list_actual_position[1]] = '.'
-                    self.nb_plays = self.nb_plays + 1
-                    return chessboard
 
 
 class Rook(Piece):
